Finish edit and fix delete

+++ public/js/app.js
@@ -3,39 +3,41 @@ jQuery(function ($) {
 
     // Delete
     $(document).on('submit', '.js-delete',  function (e) {
-        var route = $(this).attr('action'),
-            $parentRow = $(this).parents('tr'),
-            token = $(this).children('input[name="_token"]').val();
+        if (confirm('Ești sigur că vrei să ștergi?')) {
+            var route = $(this).attr('action'),
+                $parentRow = $(this).parents('tr'),
+                token = $(this).children('input[name="_token"]').val();
 
-        $.ajax({
-            url: route,
-            type: 'DELETE',
-            data: {'_token': token},
-            success: function(result) {
-                Swal.fire('Success', result.message, 'success');
-                $parentRow.fadeOut();
-            },
-            error: function() {
-                Swal.fire('Eroare', 'Ne pare rău, dar nu am putut șterge.', 'error');
-            }
-        });
+            $.ajax({
+                url: route,
+                type: 'DELETE',
+                data: {'_token': token},
+                success: function (result) {
+                    Swal.fire('Success', result.message, 'success');
+                    $parentRow.fadeOut();
+                },
+                error: function () {
+                    Swal.fire('Eroare', 'Ne pare rău, dar nu am putut șterge.', 'error');
+                }
+            });
+        }
 
         e.preventDefault();
     })
 
-    // Show form on edit
+    // Show edit form
     $(document).on('click', '.js-edit',  function (e) {
         var route = $(this).attr('href'),
             $parentRow = $(this).parents('tr');
 
         $.get(route, function(data) {
-            $parentRow.replaceWith(data).fadeIn();
+            $parentRow.hide().after(data);
         });
 
         e.preventDefault();
     })
 
-    // Submit edit form
+    // Save edit form
     $(document).on('click', '.js-save',  function (e) {
         var route = $(this).data('href'),
             $parentRow = $(this).parents('tr'),
@@ -43,11 +45,20 @@ jQuery(function ($) {
 
         $.post(route, $form.serialize(), function(html) {
             $form.fadeOut(300, function() {
-                $parentRow.hide().replaceWith(html).fadeIn(400);
+                $parentRow.prev().replaceWith(html).fadeIn(400);
+                $parentRow.remove();
             });
         });
 
         e.preventDefault();
+    })
+
+    // Cancel edit form
+    $(document).on('click', '.js-close',  function () {
+        var $parentRow = $(this).parents('tr');
+
+        $parentRow.prev().fadeIn();
+        $parentRow.remove();
     })
 
     var $changeDayPicker = $('#js-change-day-datepicker');
@@ -60,7 +71,11 @@ jQuery(function ($) {
         toggleActive: true,
         autoclose: true
     }).on('changeDate', function(e) {
-        $('#js-change-day-date').val($changeDayPicker.datepicker('getFormattedDate'));
-        $('#js-change-day-form').submit();
+        var formattedDate = $changeDayPicker.datepicker('getFormattedDate');
+
+        if(formattedDate.length !== 0) {
+            $('#js-change-day-date').val($changeDayPicker.datepicker('getFormattedDate'));
+            $('#js-change-day-form').submit();
+        }
     });
 });
